Moved up ZMQ check to provider
